third_party/blink/web_tests/virtual/force-in-order-script/wpt_internal/in-order-script-scheduling/force-in-order/no-parser-blocking-expected.txt

This is a testharness.js-based test.
[FAIL] InOrder Script Execution Order
  Uncaught Error: assert_array_equals: Execution order expected property 1 to be "Sync1" but got "EndOfBody" (expected array ["Inline1", "Sync1", "EndOfBody", "DOMContentLoaded", "Async1"] got ["Inline1", "EndOfBody", "DOMContentLoaded", "Async1", "Sync1"])
Harness: the test ran to completion.
